refactor(e2e): Split of common.ts to common and electron
fix typo
