+++ packages/suite-desktop-core/e2e/tests/bridge-tor/spawn-tor.test.ts
@@ -1,6 +1,7 @@
 import { Page } from '@playwright/test';
 
-import { launchSuite, skipFixture } from '../../support/common';
+import { skipFixture } from '../../support/common';
+import { launchSuite } from '../../support/electron';
 import { expect, test } from '../../support/fixtures';
 import { NetworkAnalyzer } from '../../support/networkAnalyzer';
